+++ praw/models/front.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 from urllib.parse import urljoin
 
 from praw.models.listing.generator import ListingGenerator
@@ -23,7 +23,7 @@ class Front(SubredditListingMixin):
         super().__init__(reddit, _data=None)
         self._path = "/"
 
-    def best(self, **generator_kwargs: str | int) -> Iterator[models.Submission]:
+    def best(self, **generator_kwargs: Any) -> Iterator[models.Submission]:
         """Return a :class:`.ListingGenerator` for best items.
 
         Additional keyword arguments are passed in the initialization of