+++ praw/models/listing/mixins/rising.py
@@ -11,13 +11,19 @@ from praw.models.listing.generator import ListingGenerator
 if TYPE_CHECKING:
     from collections.abc import Iterator
 
+    from typing_extensions import Unpack
+
     from praw import models
+    from praw.models.listing.generator import ListingGeneratorKwargs
 
 
 class RisingListingMixin(PRAWBase):
     """Mixes in the rising methods."""
 
-    def rising(self, **generator_kwargs: str | int | dict[str, str]) -> Iterator[models.Submission]:
+    if TYPE_CHECKING:
+        _path: str
+
+    def rising(self, **generator_kwargs: Unpack[ListingGeneratorKwargs]) -> Iterator[models.Submission]:
         """Return a :class:`.ListingGenerator` for rising submissions.
 
         Additional keyword arguments are passed in the initialization of