+++ praw/models/reddit/wikipage.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, cast
 
 from praw.const import API_PATH
 from praw.models.listing.generator import ListingGenerator
@@ -174,7 +174,8 @@ class WikiPage(RedditBase):
         subreddit: models.Subreddit,
         url: str,
     ) -> Iterator[dict[str, Redditor | WikiPage | str | int | bool | None]]:
-        for revision in ListingGenerator(subreddit._reddit, url, **generator_kwargs):
+        for item in ListingGenerator(subreddit._reddit, url, **generator_kwargs):
+            revision = cast("dict[str, Any]", item)
             if revision["author"] is not None:
                 revision["author"] = Redditor(subreddit._reddit, _data=revision["author"]["data"])
             revision["page"] = WikiPage(subreddit._reddit, subreddit, revision["page"], revision["id"])
@@ -228,7 +229,7 @@ class WikiPage(RedditBase):
         self.__dict__.update(data)
         super()._fetch()
 
-    def _fetch_info(self) -> tuple[str, dict[str, str], dict[str, str] | None]:
+    def _fetch_info(self) -> tuple[str, dict[str, str | models.Subreddit], dict[str, str] | None]:
         return (
             "wiki_page",
             {"subreddit": self.subreddit, "page": self.name},
@@ -288,7 +289,9 @@ class WikiPage(RedditBase):
         """
         return WikiPage(self.subreddit._reddit, self.subreddit, self.name, revision)
 
-    def revisions(self, **generator_kwargs: str | int | dict[str, str]) -> Iterator[WikiPage]:
+    def revisions(
+        self, **generator_kwargs: Any
+    ) -> Iterator[dict[str, models.Redditor | WikiPage | str | int | bool | None]]:
         """Return a :class:`.ListingGenerator` for page revisions.
 
         Additional keyword arguments are passed in the initialization of